add: package that handles plot themes
