+++ metrics-theme/src/pyosmetrics/plot_theme.py
@@ -111,3 +111,66 @@ def parse_single_issue(issue) -> dict:
         parsed_issue["days_open"] = delta.days
 
     return parsed_issue
+
+
+def count_edits_by_quarter(df):
+    """
+    Count submission edited by each editor by quarter.
+
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        Dataframe containing a ``created_at`` datetime column and
+        an ``editor`` column.
+
+    Returns
+    -------
+    pandas.DataFrame
+        DataFrame containing the number of submissions edited by each editor
+        by quarter.
+    """
+    n_edits = df.groupby(
+        by=[df["editor"], df["created_at"].dt.to_period("Q")]
+    ).count()
+    n_edits.rename(columns=dict(created_at="n_edits"), inplace=True)
+    return n_edits
+
+
+def test_count_edits_by_quarter():
+    import pandas as pd
+
+    # Build a sample df
+    sample_df = pd.DataFrame(
+        {
+            "created_at": [
+                "2024-03-01",
+                "2024-06-18",
+                "2024-11-12",
+                "2024-12-12",
+                "2024-03-11",
+                "2024-04-12",
+            ],
+            "editor": ["foo", "foo", "foo", "foo", "bar", "bar"],
+        }
+    )
+    sample_df["created_at"] = pd.to_datetime(sample_df.created_at)
+
+    # Get n_edits with the function
+    n_edits = count_edits_by_quarter(sample_df)
+
+    # Build expected df
+    expected = pd.DataFrame(
+        {
+            "n_edits": {
+                ("bar", pd.Period("2024Q1", "Q-DEC")): 1,
+                ("bar", pd.Period("2024Q2", "Q-DEC")): 1,
+                ("foo", pd.Period("2024Q1", "Q-DEC")): 1,
+                ("foo", pd.Period("2024Q2", "Q-DEC")): 1,
+                ("foo", pd.Period("2024Q4", "Q-DEC")): 2,
+            }
+        }
+    )
+    expected.index.names = ["editor", "created_at"]
+
+    # Test if output matches the expected one
+    pd.testing.assert_frame_equal(expected, n_edits)
